+++ ASDKSample/ASDKSample/RootViewController.swift
@@ -22,6 +22,12 @@ import UIKit
 import TinkoffASDKCore
 import TinkoffASDKUI
 
+extension AddCardOptions {
+    static let cardOptions = AddCardOptions(
+        attachCardData: AdditionalData(data: ["/AttachKey": "/AttachValue"])
+    )
+}
+
 struct Product: Codable {
 
     var price: NSDecimalNumber
@@ -159,6 +165,7 @@ class RootViewController: UITableViewController {
             sdk.presentCardList(
                 on: self,
                 customerKey: AppSetting.shared.activeSdkCredentials.customerKey,
+                addCardOptions: .cardOptions,
                 cardScannerDelegate: self
             )
         }
@@ -172,7 +179,12 @@ class RootViewController: UITableViewController {
 
             let customerKey = AppSetting.shared.activeSdkCredentials.customerKey
 
-            sdk.presentAddCard(on: self, customerKey: customerKey, cardScannerDelegate: nil) { [weak self] result in
+            sdk.presentAddCard(
+                on: self,
+                customerKey: customerKey,
+                addCardOptions: .cardOptions,
+                cardScannerDelegate: nil
+            ) { [weak self] result in
                 self?.addingNewCardCompleted(result: result)
             }
         }