+++ TinkoffASDKUI/TinkoffASDKUITests/Infrastructure/Mocks/CardsControllerMock.swift
@@ -21,14 +21,14 @@ final class CardsControllerMock: ICardsController {
 
     // MARK: - addCard
 
-    typealias AddCardArguments = (options: CardOptions, completion: (AddCardResult) -> Void)
+    typealias AddCardArguments = (cardData: CardData completion: (AddCardResult) -> Void)
 
     var addCardCallsCount = 0
     var addCardReceivedArguments: AddCardArguments?
     var addCardReceivedInvocations: [AddCardArguments?] = []
     var addCardCompletionClosureInput: AddCardResult?
 
-    func addCard(options: CardOptions, completion: @escaping (AddCardResult) -> Void) {
+    func addCard(cardData: CardData completion: @escaping (AddCardResult) -> Void) {
         addCardCallsCount += 1
         let arguments = (options, completion)
         addCardReceivedArguments = arguments